+++ examples/user_assigned_identity/main.tf
@@ -10,6 +10,13 @@ module "rg" {
   location = "North Europe"
 }
 
+resource "azurerm_user_assigned_identity" "aks_identity" {
+  resource_group_name = module.rg.name
+  location            = module.rg.location
+
+  name = "aks_identity"
+}
+
 module "aks" {
   source = "../.."
 
@@ -18,6 +25,8 @@ module "aks" {
   dns_prefix          = "demolab"
 
   default_pool_name = "default"
+
+  user_assigned_identity_id = azurerm_user_assigned_identity.aks_identity.id
 
   depends_on = [module.rg]
 }