+++ CHANGELOG.md
@@ -3,14 +3,32 @@
 ## [Unreleased]
 
 
+<a name="0.3.0"></a>
+## [0.3.0] - 2021-02-26
+### Feat
+- network profile
+- active directory integration
+- add more options
+- aks add ons
+- basic cluster
+- basic cluster
+
+### Fix
+- documentation
+
+
 <a name="0.2.0"></a>
-## [0.2.0] - 2021-02-24
+## [0.2.0] - 2021-02-25
 ### Feat
 - ssh-key module
+
+### Pull Requests
+- Merge pull request [#1](https://github.com/bcochofel/terraform-azurerm-aks/issues/1) from bcochofel/first_release
 
 
 <a name="0.1.0"></a>
 ## 0.1.0 - 2021-02-23
 
-[Unreleased]: https://github.com/bcochofel/terraform-azurerm-aks/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/bcochofel/terraform-azurerm-aks/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/bcochofel/terraform-azurerm-aks/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/bcochofel/terraform-azurerm-aks/compare/0.1.0...0.2.0